chore: add tests
