+++ test/blocks/header/header.test.js
@@ -1,12 +1,12 @@
 /* eslint-disable no-unused-expressions */
 /* global describe it */
 
-import { readFile } from '@web/test-runner-commands';
+import { readFile, setViewport } from '@web/test-runner-commands';
 import { expect } from '@esm-bundle/chai';
 
 document.body.innerHTML = await readFile({ path: '../../scripts/dummy.html' });
 
-const { buildBlock, decorateBlock, loadBlock } = await import('../../../scripts/scripts.js');
+const { buildBlock, decorateBlock, loadBlock } = await import('../../../scripts/lib-franklin.js');
 
 document.body.innerHTML = await readFile({ path: '../../scripts/body.html' });
 
@@ -34,7 +34,8 @@ describe('Header block', () => {
     expect(nav.getAttribute('aria-expanded')).to.equal('false');
   });
 
-  it('Section title shows and hides section', async () => {
+  it('Section title shows and hides section on desktop', async () => {
+    await setViewport({ width: 900, height: 640 });
     const sections = document.querySelector('.header .nav-sections');
     const title = sections.querySelector(':scope li');
     title.click();
