+++ test/scripts/body.html
@@ -45,7 +45,7 @@
 &lt;script src="/scripts.js" type="module"&gt;&lt;/script&gt;
 &lt;link rel="stylesheet" href="/styles.css"/&gt;
         </code></pre>
-        <div class="example"><
+        <div class="hero">
             <p>Lorem ipsum dolor sit amet, <strong>consectetur</strong> adipiscing elit, sed do eiusmod tempor incididunt ut
                 labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut
                 aliquip ex ea commodo <em>consequat</em>. Duis aute irure dolor in reprehenderit in voluptate velit esse
