+++ test/scripts/types.d.ts
@@ -1,0 +1,5 @@
+import type * as ScriptsModule from '../../scripts/scripts';
+import type * as LibModule from '../../scripts/lib-franklin';
+
+export type Scripts = typeof ScriptsModule;
+export type LibFranklin = typeof LibModule;
